# swagger.yaml
openapi: 3.0.0
info:
  title: API de Usuarios
  version: 1.0.0
  description: API para la gestión de usuarios. Permite obtener todos los usuarios y crear nuevos usuarios.
paths:
  /users:
    get:
      summary: Obtener todos los usuarios
      description: Recupera una lista de todos los usuarios almacenados en la base de datos.
      responses:
        '200':
          description: Lista de usuarios obtenida correctamente.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: ID único del usuario.
                    nombre:
                      type: string
                      description: Nombre del usuario.
                    email:
                      type: string
                      description: Correo electrónico del usuario.
        '500':
          description: Error interno del servidor. Ocurrió un problema al obtener los usuarios.
          
    post:
      summary: Crear un nuevo usuario
      description: Crea un nuevo usuario con los datos proporcionados en la solicitud.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                  description: Nombre del usuario.
                email:
                  type: string
                  description: Correo electrónico del usuario.
              required:
                - nombre
                - email
      responses:
        '201':
          description: Usuario creado exitosamente. Se devuelve el usuario creado con su ID.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: ID único del usuario creado.
                  nombre:
                    type: string
                    description: Nombre del usuario creado.
                  email:
                    type: string
                    description: Correo electrónico del usuario creado.
        '400':
          description: Datos de entrada inválidos. El nombre y el correo son obligatorios y deben ser válidos.
        '409':
          description: El usuario ya existe. Ya hay un usuario con el mismo correo electrónico en la base de datos.
        '500':
          description: Error interno del servidor. Ocurrió un problema al crear el usuario.